Safari-611.1.10/LayoutTests/fast/dom/Window/no-window-resize-on-document-size-change-expected.txt

Test that changing the viewport causes a window.resize event.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS windowResizeCount is 0
PASS successfullyParsed is true

TEST COMPLETE
